feat(dsv4): migrate weight loading to framework descriptor path
Replaces the hand-rolled DSv4 safetensors loader with the framework's
``WeightInfo`` / ``fastsafetensors`` / per-quant subclass / postprocess
pipeline, then cleans up the V4Transformer factory-mode bind sites so
each module consumes ``ModelWeights`` tensors directly.

What's in this commit
=====================

1. **Framework descriptor declarations** (``rtp_llm/models/deepseek_v4.py``):
   declare every per-layer + global ``WeightModule`` the loader needs
   to populate ``ModelWeights`` from a V4-Flash safetensors checkpoint
   - attention norms / sink, dense MQA FP8 linears (auto-wrapped by
   ``V4PerBlockFp8Weight``), CSA/HCA compressors (outer + inner
   indexer variants), lightning indexer, mHC residual scales, FFN
   norm, hash/noaux_tc router, shared expert FP8, FP4 routed experts
   (per-expert int8-packed weight + UE8M0 group-32 scale).  Per-layer
   module set is heterogeneous, driven by ``compress_ratios`` and
   ``num_hash_layers``.  MTP weight class declared (currently gated
   behind ``n_mtp_layers=0``).

2. **FP4 quant subclass** (``per_group_fp4_quant_weight.py``):
   V4PerGroupFp4Weight ``support()`` scoped to V4 namespace; binds
   ``W.v4_routed_w*_w/s`` as ordinary stacked MoE tensors with
   ``data_type=torch.int8`` (matches I8 ckpt dtype so DeepGEMM's
   ``kPackedFP4`` path fires) and ``data_type=torch.float8_e8m0fnu``
   for scales.  No postprocess transform — V4 ckpt layout is already
   DeepGEMM-consumable.

3. **FP8 quant subclass extension** (``per_block_fp8_quant_weight.py``):
   ``V4PerBlockFp8Weight`` whitelist + ``_v4_build_pair`` to derive
   the .scale ckpt key from .weight; SM100+ ``requant_weight_ue8m0``
   path stays in shared base postprocess (produces TMA-aligned
   packed int32 scales DeepGEMM E8M0 fp8_gemm_nt requires).

4. **Loader-side fixes** that the migration depends on:
   - ``utils/torch_patch.py``: patch ``dist.broadcast`` to handle
     UE8M0 (``float8_e8m0fnu``) ckpt scales (PyTorch dist doesn't
     know the dtype natively; broadcast as raw bytes).
   - ``utils/database.py``: minor accommodations for the new dtype.
   - ``utils/model_weight.py``: V4 W keys + identity helper extensions.

5. **Bridge adapter** (``deepseek_v4_model.py::_flatten_framework_weights``):
   translates ``ModelWeights[layer][W.v4_*]`` + globals back to flat
   ckpt-style keys for V4Transformer factory mode.  Uses ``pop``
   (not ``get``) for both globals and per-layer non-routed-expert
   keys so storage is released as soon as the module binds the
   tensor — without this, every weight has TWO Python owners
   (ModelWeights dict + module attr) and ~80 GB of attention/shared/
   global weights are doubly held forever (V4-Flash 64k prefill OOM).
   Routed-expert stacks unstacked into per-expert keys via the
   second pass.  Unknown W keys logged loudly so future descriptor
   gaps surface immediately rather than silently dropping tensors.

6. **Mega MoE setup HBM serialisation** (``moe.py::_setup_mega_moe``):
   serialise L1 -> L2 -> mega-transform with ``del`` +
   ``torch.cuda.empty_cache()`` between stages so the live set is
   bounded by a single (w13, s13_fp32) or (w2, s2_fp32) stack.
   Without this V4-Pro cp4 OOMs around L15 (per-layer +3.13 GB net
   transient peak).

7. **Module-side bind cleanup**: drop ``nn.Parameter`` wrapping +
   placeholder holder classes (``_LinearHolder``, ``_EmbeddingHolder``,
   ``_NormHolder``, ``_LMHead``, factory-mode ``QuantizedLinear``
   for wo_a) across attention / block / compressor / indexer / mhc
   / moe / qlinear / transformer.  Each module now binds the loader's
   tensor reference directly.  ``_RMSNorm`` / ``_CompressorNorm``
   accept ``weight=`` at construction (matches framework
   ``RMSResNorm(weights[W.pre_ln_gamma], eps=...)`` pattern).
   Drop 19 no-op ``.to(bf16)`` / ``.float()`` casts at bind sites
   that match the loader's already-cast dtype; align W descriptors
   with consumer dtype (``W.v4_router_tid2eid`` int64 -> int32,
   ``W.v4_compressor_wkv/wgate`` + inner indexer variants gain
   ``data_type=torch.float32`` so the loader delivers fp32 directly).
   Real bf16 -> fp32 cast at bind retained for ``head.weight``
   (descriptor stays bf16 to halve NCCL broadcast bandwidth).
   ``device`` derivation switches to ``self.embed.weight.device``
   (V4Transformer no longer has parameters).

8. **Legacy loader removal** (``modules/dsv4/weight_loader.py``):
   delete the bypass safetensors loader + its single-layer / full-model
   smoke harness scripts + DSV4_USE_LEGACY_LOADER escape hatch.

End-to-end validation
=====================

* V4-Flash tp1 smoke (``v4_flash_native_fp4_fp8_tp1_b200``): byte-identical
  outputs vs baseline (verified by stash + rerun: same actual outputs,
  same pre-existing 2+2 golden drift on both paths).
* V4-Pro cp4 EP4: per-layer net delta drops from +3.13 GB to ~+0.1 GB,
  total post-load HBM 224.4 GB / 268 GB, all 61 layers + first prefill
  complete successfully.

+++ rtp_llm/models_py/kernels/cuda/fp8_kernel/fp8_kernel.py
@@ -322,7 +322,7 @@ def block_quant_dequant(
     )
     x_scale_repeat = x_scale_repeat[..., :n, :k]
 
-    return (x_q_block.to(torch.float32) * x_scale_repeat).to(dtype)
+    return (x_q_block.to(torch.float32) * x_scale_repeat.to(torch.float32)).to(dtype)
 
 
 # COPIED FROM DeepGEMM
